+++ rcsbsearchapi/search.py
@@ -63,6 +63,8 @@ ChemSimType = Literal["formula", "descriptor"]  # possible query types for chemi
 ChemSimMatchType = Literal["graph-relaxed-stereo", "graph-relaxed", "fingerprint-similarity",  # possible match types for descriptor query type (Chemical similarity search)
                            "sub-struct-graph-relaxed-stereo", "sub-struct-graph-relaxed", "graph-exact"]
 TAndOr = Literal["and", "or"]
+VerbosityLevel = Literal["compact", "minimal", "verbose"]
+AggregationType = Literal["terms", "histogram", "date_histogram", "range", "date_range", "cardinality"]
 # All valid types for Terminal values
 TValue = Union[
     str,
@@ -170,21 +172,40 @@ class Query(ABC):
         """Symmetric difference: `a ^ b`"""
         return (self & ~other) | (~self & other)
 
-    def exec(self, return_type: ReturnType = "entry", rows: int = 10000, return_content_type: List[ReturnContentType] = ["experimental"]) -> "Session":
+    def exec(
+        self,
+        return_type: ReturnType = "entry",
+        rows: int = 10000,
+        return_content_type: List[ReturnContentType] = ["experimental"],
+        results_verbosity: VerbosityLevel = "compact"
+    ) -> "Session":
         # pylint: disable=dangerous-default-value
         """Evaluate this query and return an iterator of all result IDs"""
-        return Session(self, return_type, rows, return_content_type)
+        return Session(self, return_type, rows, return_content_type, results_verbosity)
 
-    def __call__(self, return_type: ReturnType = "entry", rows: int = 10000, return_content_type: List[ReturnContentType] = ["experimental"]) -> "Session":
+    def __call__(
+        self,
+        return_type: ReturnType = "entry",
+        rows: int = 10000,
+        return_content_type: List[ReturnContentType] = ["experimental"],
+        results_verbosity: VerbosityLevel = "compact"
+    ) -> "Session":
         # pylint: disable=dangerous-default-value
         """Evaluate this query and return an iterator of all result IDs"""
-        return self.exec(return_type, rows, return_content_type)
+        return self.exec(return_type, rows, return_content_type, results_verbosity)
 
     def count(self, return_type: ReturnType = "entry", return_content_type: List[ReturnContentType] = ["experimental"]) -> int:
         # pylint: disable=dangerous-default-value
+        """Get the number of results found by this query"""
         s = Session(self, return_type, 0, return_content_type)
         response = s._single_query()
         return response["total_count"] if response else 0
+
+    def facets(self, return_type: ReturnType = "entry", facets: Union["Facet", "FilterFacet", List[Union["Facet", "FilterFacet"]]] = None) -> List:
+        """Perform a facets query and return the buckets"""
+        s = Session(self, return_type=return_type, rows=0, facets=facets)
+        response = s._single_query()
+        return response["facets"] if response else []
 
     @overload
     def and_(self, other: "Query") -> "Query":
@@ -1224,6 +1245,161 @@ class Value(Generic[T]):
         return attr.less_or_equal(self.value)
 
 
+@dataclass(frozen=True)
+class Range:
+    """Primarily for use with "range" and "date_range" aggregations with the Facet class.
+    include_upper and include_lower should not be used with Facet queries."""
+
+    start: Union[str, float] = None
+    end: Union[str, float] = None
+    include_lower: Optional[bool] = None
+    include_upper: Optional[bool] = None
+
+    def to_dict(self) -> dict:
+        d = {}
+        if self.start is not None:
+            d["from"] = self.start
+        if self.end is not None:
+            d["to"] = self.end
+        if self.include_lower is not None:
+            d["include_lower"] = self.include_lower
+        if self.include_upper is not None:
+            d["include_upper"] = self.include_upper
+        return d
+
+
+class Facet:
+    """Facets can be used (in conjunction with the facet() function on a Query) in order to group and perform calculations and statistics on PDB data.
+    Facets arrange search results into categories (buckets) based on the requested field values.
+    """
+    def __init__(
+        self,
+        name: str,
+        aggregation_type: AggregationType,
+        attribute: str,
+        interval: Optional[Union[int, str]] = None,
+        ranges: Optional[List[Range]] = None,
+        min_interval_population: Optional[int] = None,
+        max_num_intervals: Optional[int] = None,
+        precision_threshold: Optional[int] = None,
+        nested_facets: Optional[Union["Facet", "FilterFacet", List[Union["Facet", "FilterFacet"]]]] = None
+    ):
+        """Initialize Facet object for use in a faceted query.
+
+        Args:
+            name (str): Specifies the name of the aggregation.
+            aggregation_type (AggregationType): Specifies the type of the aggregation. Can be "terms", "histogram", "date_histogram", "range", "date_range", or "cardinality".
+            attribute (str): Specifies the full attribute name to aggregate on.
+            interval (Optional[Union[int, str]], optional): Size of the intervals into which a given set of values is divided. Required only for use with
+            "histogram" and "date_histogram" aggregation types (defaults to None if not included).
+            ranges (Optional[List[Range]], optional): A set of ranges, each representing a bucket. Note that this aggregation includes the 'from' value and
+            excludes the 'to' value for each range. Should be a list of Range objects (leave the "include_lower" and "include_upper" fields empty). Required
+            only for use with "range" and "date_range" aggregation types (defaults to None if not included).
+            min_interval_population (Optional[int], optional): Minimum number of items (>= 0) in the bin required for the bin to be returned. Only for use with
+            "terms", "histogram", and "date_histogram" facets (defaults to 1 for these aggregation types, otherwise defaults to None).
+            max_num_intervals (Optional[int], optional): Maximum number of intervals (<= 65336) to return for a given facet. Only for use with "terms"
+            aggregation type (defaults to 65336 for this aggregation type, otherwise defaults to None).
+            precision_threshold (Optional[int], optional): Allows to trade memory for accuracy, and defines a unique count (<= 40000) below which counts are
+            expected to be close to accurate. Only for use with "cardinality" aggregation type (defaults to 40000 for this aggregation type, otherwise defaults to None).
+            nested_facets (Optional[Union[&quot;Facet&quot;, &quot;FilterFacet&quot;, List[Union[&quot;Facet&quot;, &quot;FilterFacet&quot;]]]], optional): Enables
+            multi-dimensional aggregations. Should contain a List of Facets or FilterFacets. Can be used with any aggregation type. Defaults to None.
+        """
+        self.name = name
+        self.aggregation_type = aggregation_type
+        self.attribute = attribute
+        self.interval = interval
+        self.ranges = ranges
+        self.min_interval_population = min_interval_population
+        self.max_num_intervals = max_num_intervals
+        self.precision_threshold = precision_threshold
+        self.nested_facets = nested_facets if isinstance(nested_facets, list) or nested_facets is None else [nested_facets]
+
+        if self.aggregation_type == "terms":
+            if self.min_interval_population is None:
+                self.min_interval_population = 1
+            if self.max_num_intervals is None:
+                self.max_num_intervals = 65536
+        elif self.aggregation_type == "histogram":
+            if self.min_interval_population is None:
+                self.min_interval_population = 1
+        elif self.aggregation_type == "date_histogram":
+            if self.min_interval_population is None:
+                self.min_interval_population = 1
+        elif self.aggregation_type == "cardinality":
+            if self.precision_threshold is None:
+                self.precision_threshold = 40000
+
+    def to_dict(self) -> dict:
+        facet_dict = dict(name=self.name, aggregation_type=self.aggregation_type, attribute=self.attribute)
+        if self.interval is not None:
+            facet_dict["interval"] = self.interval
+        if self.ranges is not None:
+            facet_dict["ranges"] = [r.to_dict() for r in self.ranges]
+        if self.min_interval_population is not None:
+            facet_dict["min_interval_population"] = self.min_interval_population
+        if self.max_num_intervals is not None:
+            facet_dict["max_num_intervals"] = self.max_num_intervals
+        if self.precision_threshold is not None:
+            facet_dict["precision_threshold"] = self.precision_threshold
+        if self.nested_facets is not None:
+            facet_dict["facets"] = [f.to_dict() for f in self.nested_facets if f is not None]
+        return facet_dict
+
+
+class TerminalFilter:
+    """
+    Terminal filter class for use with FilterFacet queries
+    """
+
+    def __init__(
+        self,
+        attribute: str,
+        operator: Literal["equals", "greater", "greater_or_equal", "less", "less_or_equal", "range", "exact_match", "in", "exists"],
+        value: Optional[Union[str, int, float, bool, Range, List[str], List[int], List[float]]] = None,
+        negation: bool = False,
+        case_sensitive: bool = False
+    ):
+        self.attribute = attribute
+        self.operator = operator
+        self.value = value
+        self.negation = negation
+        self.case_sensitive = case_sensitive
+
+    def to_dict(self):
+        tf_dict = dict(type="terminal", service="text", parameters=dict(attribute=self.attribute, operator=self.operator, negation=self.negation, case_sensitive=self.case_sensitive))
+        if self.value is not None:
+            tf_dict["parameters"]["value"] = self.value
+        return tf_dict
+
+
+class GroupFilter:
+    """
+    Group filter class for use with FilterFacet queries
+    """
+
+    def __init__(self, logical_operator: TAndOr, nodes: List[Union["TerminalFilter", "GroupFilter"]]):
+        self.logical_operator = logical_operator,
+        self.nodes = nodes
+
+    def to_dict(self):
+        return dict(type="group", logical_operator=self.logical_operator[0], nodes=[node.to_dict() for node in self.nodes])
+
+
+class FilterFacet:
+    """Facet queries using Filters"""
+
+    def __init__(
+        self,
+        filters: Union[TerminalFilter, GroupFilter],
+        facets: Union[Facet, "FilterFacet", List[Union[Facet, "FilterFacet"]]]
+    ):
+        self.filter = filters
+        self.facets = facets if isinstance(facets, list) else [facets]
+
+    def to_dict(self):
+        return dict(filter=self.filter.to_dict(), facets=[facet.to_dict() for facet in self.facets])
+
+
 class Session(Iterable[str]):
     """A single query session.
 
@@ -1237,10 +1413,18 @@ class Session(Iterable[str]):
     start: int
     rows: int
     return_content_type: List[ReturnContentType]
+    results_verbosity: VerbosityLevel
+    facets: Optional[Union[Facet, FilterFacet, List[Union[Facet, FilterFacet]]]]
 
     def __init__(
         # parameter added below for computed model inclusion
-        self, query: Query, return_type: ReturnType = "entry", rows: int = 10000, return_content_type: List[ReturnContentType] = ["experimental"]
+        self,
+        query: Query,
+        return_type: ReturnType = "entry",
+        rows: int = 10000,
+        return_content_type: List[ReturnContentType] = ["experimental"],
+        results_verbosity: VerbosityLevel = "compact",
+        facets: Optional[Union[Facet, FilterFacet, List[Union[Facet, FilterFacet]]]] = None
         # pylint: disable=dangerous-default-value
     ):
         self.query_id = Session.make_uuid()
@@ -1249,6 +1433,8 @@ class Session(Iterable[str]):
         self.start = 0
         self.rows = rows
         self.return_content_type = return_content_type
+        self.results_verbosity = results_verbosity
+        self.facets = facets
 
     @staticmethod
     def make_uuid() -> str:
@@ -1268,13 +1454,19 @@ class Session(Iterable[str]):
 
     def _make_params(self, start=0):
         "Generate GET parameters as a dict"
-        return dict(
+        query_dict = dict(
             query=self.query.to_dict(),
             return_type=self.return_type,
             request_info=dict(query_id=self.query_id, src="ui"),  # "TODO" src deprecated?
             # v1 -> v2: pager parameter is renamed to paginate and results_content_type parameter added (which has a list as its value)
-            request_options=dict(paginate=dict(start=start, rows=self.rows), results_content_type=self.return_content_type),
+            request_options=dict(paginate=dict(start=start, rows=self.rows), results_content_type=self.return_content_type, results_verbosity=self.results_verbosity),
         )
+        if self.facets is not None:
+            if type(self.facets) is list:
+                query_dict["request_options"]["facets"] = [facet.to_dict() for facet in self.facets]
+            else:
+                query_dict["request_options"]["facets"] = [self.facets.to_dict()]
+        return query_dict
 
     def _single_query(self, start=0) -> Optional[Dict]:
         "Fires a single query"
@@ -1300,27 +1492,27 @@ class Session(Iterable[str]):
         response = self._single_query(start=start)
         if response is None:
             return  # be explicit for mypy
-        identifiers = self._extract_identifiers(response)
+        result_set = response["result_set"] if response else []
         start += self.rows
-        logging.debug("Got %s ids", len(identifiers))
+        logging.debug("Got %s ids", len(result_set))
 
-        if len(identifiers) == 0:
+        if len(result_set) == 0:
             return
-        yield from identifiers
+        yield from result_set
 
         total = response["total_count"]
 
         while start < total:
-            assert len(identifiers) == self.rows
+            assert len(result_set) == self.rows
             req_count += 1
             if req_count == REQUESTS_PER_SECOND:
                 time.sleep(1.2)  # This prevents the user from bottlenecking the server with requests.
                 req_count = 0
             response = self._single_query(start=start)
-            identifiers = self._extract_identifiers(response)
-            logging.debug("Got %s ids", len(identifiers))
+            result_set = response["result_set"] if response else []
+            logging.debug("Got %s ids", len(result_set))
             start += self.rows
-            yield from identifiers
+            yield from result_set
 
     def iquery(self, limit: Optional[int] = None) -> List[str]:
         """Evaluate the query and display an interactive progress bar.
@@ -1333,18 +1525,18 @@ class Session(Iterable[str]):
         if response is None:
             return []
         total = response["total_count"]
-        identifiers = self._extract_identifiers(response)
-        if limit is not None and len(identifiers) >= limit:
-            return identifiers[:limit]
+        result_set = response["result_set"] if response else []
+        if limit is not None and len(result_set) >= limit:
+            return result_set[:limit]
 
         pages = math.ceil((total if limit is None else min(total, limit)) / self.rows)
 
         for page in trange(1, pages, initial=1, total=pages):
             response = self._single_query(page * self.rows)
-            ids = self._extract_identifiers(response)
-            identifiers.extend(ids)
+            next_results = response["result_set"] if response else []
+            result_set.extend(next_results)
 
-        return identifiers[:limit]
+        return result_set[:limit]
 
     def rcsb_query_editor_url(self) -> str:
         """URL to edit this query in the RCSB PDB query editor"""