+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+## v1.5.0 (2024-02-07)
+
+- Added support for faceted queries
+- Added support for retrieving the scores of results (via `results_verbosity`)
+
 ## v1.4.2 (2023-10-17)
 
-- Add support for count queries
+- Added support for count queries
 
 ## v1.4.1 (2023-09-14)